+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.8
 - Added setlinux, setmacos, setwindows
+- Set Scriban default newline to "\n"
 
 ## 1.0.7
 - Added Newtonsoft.Json